[#11] 기본 엔티티 정의

+++ src/db/entities/task_definition.rs
@@ -4,17 +4,17 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "task_definition")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
-    pub name: String,
-    pub version: String,
+    pub id: u64, // primary key
+    pub name: String,    // task name
+    pub version: String, // task version
 
-    pub image: String,
-    pub command: Option<String>,
-    pub args: Option<String>,
-    pub env: Option<String>,
+    pub image: String,           // docker image
+    pub command: Option<String>, // docker run command
+    pub args: Option<String>,    // docker run arguments
+    pub env: Option<String>,     // environment variables
 
-    pub memory_limit: Option<u32>,
-    pub cpu_limit: Option<u32>,
+    pub memory_limit: Option<u32>, // memory limit in MB
+    pub cpu_limit: Option<u32>,    // cpu limit (default 1024)
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
